requirements.txt
astroid==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
platformdirs==2.4.0
psycopg2==2.9.1
pylint==2.11.1
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dotenv==0.19.0
requests==2.26.0
SQLAlchemy==1.4.25
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
